[fix] Refactored register flow (#64)
