+++ app/(protected)/(tabs)/_layout.tsx
@@ -1,7 +1,7 @@
 import { Tabs } from "expo-router";
 import { Heart, HomeIcon, MessageCircle, PieChart, Settings } from "lucide-react-native";
 import { Platform } from "react-native";
-import { useTheme } from "../theme/context";
+import { useTheme } from "../../theme/context";
 
 export default function TabLayout() {
     const { theme } = useTheme();
