+++ app/auth/register/_layout.tsx
@@ -5,7 +5,6 @@ export default function RegisterLayout() {
         <Stack>
             <Stack.Screen name="index" options={{ headerShown: false }} />
             <Stack.Screen name="withEmail" options={{ headerShown: false }} />
-            <Stack.Screen name="withPhoneNumber" options={{ headerShown: false }} />
             <Stack.Screen name="sendVerificationEmail" options={{ headerShown: false }} />
             <Stack.Screen name="sendVerificationPhone" options={{ headerShown: false }} />
             <Stack.Screen name="gender" options={{ headerShown: false }} />
