+++ app/theme/context.tsx
@@ -2,6 +2,7 @@ import { createContext, useContext } from "react";
 import { makeGlobalStyles } from "@/app/theme/styles";
 import { colors } from "@/app/theme/theme";
 
+// Made with https://claude.ai
 type TThemeContext = {
     theme: typeof colors.light;
     gs: ReturnType<typeof makeGlobalStyles>;
